backend/requirements.txt
Django>=5.0,<5.1
django-ninja>=1.0
django-cors-headers>=4.3
django-environ>=0.11
cryptography>=41.0
Pillow>=10.0
django-extensions>=3.2
python-dotenv>=1.0
gunicorn>=21.2
whitenoise>=6.6
djangorestframework>=3.14
PyJWT>=2.8.0  # For JWT decoding and verification
requests>=2.31.0  # For fetching OIDC config and JWKS
python-dateutil>=2.8
pytz>=2023.3
# sentry-sdk>=1.32
psycopg2-binary>=2.9
redis>=4.5
dj-database-url>=2.1
inflect>=7.0
pydantic[email]>=2.0  # For email validation in Django Ninja schemas